made SettingList and SettingDict as an uix widget
