+++ designer/buildozer_spec_editor.py
@@ -1,22 +1,16 @@
 import os
 import json
-import re
 import tempfile
 import webbrowser
 import designer
 
 from kivy.app import App
-from kivy.core.window import Window
-from kivy.metrics import dp
+
 from kivy.uix.boxlayout import BoxLayout
-from kivy.uix.button import Button
-from kivy.uix.popup import Popup
-from kivy.uix.togglebutton import ToggleButton
-from kivy.uix.widget import Widget
-from kivy.properties import ObjectProperty, ConfigParser, DictProperty, \
-    ListProperty, StringProperty, BooleanProperty
+from kivy.properties import ObjectProperty, ConfigParser, StringProperty
 from kivy.uix.settings import Settings, InterfaceWithSidebar, MenuSidebar,\
-    ContentPanel, SettingsPanel, SettingItem, SettingSpacer
+    ContentPanel, SettingsPanel
+from designer.uix.settings import SettingList, SettingDict
 
 
 class SpecContentPanel(ContentPanel):
@@ -69,239 +63,6 @@ class SpecSettingsPanel(SettingsPanel):
         if not value and not self.config.has_option(section, key):
             return False
         super(SpecSettingsPanel, self).set_value(section, key, value)
-
-
-class SpecSettingDict(SettingItem):
-    '''Implementation of an option list on top of a :class:`SettingItem`.
-    Based on SettingOptions, but implemented to use DictProperty.
-    It is visualized with a :class:`~kivy.uix.label.Label` widget that, when
-    clicked, will open a :class:`~kivy.uix.popup.Popup` with a
-    list of options from which the user can select.
-    '''
-
-    options = DictProperty({})
-    '''Dict with keys to be saved and visible values to the user
-
-    :attr:`options` is a :class:`~kivy.properties.DictProperty` and defaults
-    to {}.
-    '''
-
-    popup = ObjectProperty(None, allownone=True)
-    '''(internal) Used to store the current popup when it is shown.
-
-    :attr:`popup` is an :class:`~kivy.properties.ObjectProperty` and defaults
-    to None.
-    '''
-
-    def on_panel(self, instance, value):
-        if value is None:
-            return
-        self.bind(on_release=self._create_popup)
-
-    def _set_option(self, instance):
-        self.value = instance.key
-        self.popup.dismiss()
-
-    def _create_popup(self, instance):
-        # create the popup
-        content = BoxLayout(orientation='vertical', spacing='5dp')
-        popup_width = min(0.95 * Window.width, dp(500))
-        self.popup = popup = Popup(
-            content=content, title=self.title, size_hint=(None, None),
-            size=(popup_width, '400dp'))
-        popup.height = len(self.options) * dp(55) + dp(150)
-
-        # add all the options
-        content.add_widget(Widget(size_hint_y=None, height=1))
-        uid = str(self.uid)
-        for key in self.options:
-            state = 'down' if key == self.value else 'normal'
-            btn = ToggleButton(text=self.options[key], state=state, group=uid)
-            btn.key = key
-            btn.bind(on_release=self._set_option)
-            content.add_widget(btn)
-
-        # finally, add a cancel button to return on the previous panel
-        content.add_widget(SettingSpacer())
-        btn = Button(text='Cancel', size_hint_y=None, height=dp(50))
-        btn.bind(on_release=popup.dismiss)
-        content.add_widget(btn)
-
-        # and open the popup !
-        popup.open()
-
-
-class SettingListCheckItem(BoxLayout):
-    '''Widget with a check button and a label to display each item of list
-    '''
-    item_text = StringProperty('')
-    '''Item text. :attr:`item_text` is a
-    :class:`~kivy.properties.ObjectProperty` and defaults to ''
-    '''
-
-    active = BooleanProperty(False)
-    '''Alias to the checkbox active state.
-    :attr:`checked` is a :class:`~kivy.properties.BooleanProperty` and
-    defaults to False
-    '''
-
-    def __init__(self, **kwargs):
-        super(SettingListCheckItem, self).__init__(**kwargs)
-
-    def on_active(self, instance, *args):
-        '''Callback to update the active value
-        '''
-        self.active = instance.active
-
-
-class SettingListContent(BoxLayout):
-    '''Widget to display SettingList
-    '''
-    setting = ObjectProperty(None)
-    '''(internal) Reference to the setting SpecSettingList.
-    :attr:`setting` is a :class:`~kivy.properties.ObjectProperty` and
-    defaults to None
-    '''
-
-    custom_item_layout = ObjectProperty(None)
-    '''(internal) Widget that allows enter a custom item to the list.
-    :attr:`custom_item` is a :class:`~kivy.properties.ObjectProperty` and
-    defaults to None
-    '''
-
-    txt_custom_item = ObjectProperty(None)
-    '''(internal) TextInput with the custom item name.
-    :attr:`txt_custom_item` is a :class:`~kivy.properties.ObjectProperty` and
-    defaults to None
-    '''
-
-    item_list = ObjectProperty(None)
-    '''(internal) Widget that shows all items in a list.
-    :attr:`item_list` is a :class:`~kivy.properties.ObjectProperty` and
-    defaults to None
-    '''
-
-    selected_items = ListProperty([])
-    '''List of selected items names. Updated only after clicking on Apply button
-    :attr:`selected_item` is a :class:`~kivy.properties.ListProperty` and
-    defaults to []
-    '''
-
-    __events__ = ('on_apply', 'on_cancel', )
-
-    def __init__(self, **kwargs):
-        super(SettingListContent, self).__init__(**kwargs)
-        if not self.setting.allow_custom:
-            self.remove_widget(self.custom_item_layout)
-
-    def show_items(self, *args):
-        '''Update the list of items
-        '''
-        self.clear_items()
-        self.setting.items.sort()
-        for item in self.setting.items:
-            i = SettingListCheckItem(item_text=item)
-            if item in self.selected_items:
-                i.active = True
-            self.item_list.add_widget(i)
-
-    def clear_items(self, *args):
-        '''Remove all items from the item_list
-        '''
-        self.item_list.clear_widgets()
-
-    def on_apply_pressed(self, *args):
-        '''Event handler to Apply button.
-        Get selected items and update the selected_items.
-        '''
-        self.update_selected_list()
-        self.dispatch('on_apply', self.selected_items)
-
-    def update_selected_list(self, *args):
-        '''Update selected_items with the selected items
-        '''
-        self.selected_items = []
-        for child in self.item_list.children:
-            if child.active:
-                self.selected_items.append(str(child.item_text))
-
-    def add_custom_item(self, *args):
-        '''Add a custom item to the list
-        '''
-        txt = self.txt_custom_item.text.strip()
-        self.txt_custom_item.text = ''
-        if txt and txt not in self.setting.items:
-            self.setting.items.append(txt)
-        self.update_selected_list()
-        self.show_items()
-
-    def on_cancel(self, *args):
-        '''Event handler to Cancel button.
-        '''
-        pass
-
-    def on_apply(self, *args):
-        '''Event handler to Apply button
-        '''
-        pass
-
-
-class SpecSettingList(SettingItem):
-    '''Implementation of an multi selection list on top of :class:`SettingItem`.
-    '''
-
-    items = ListProperty([])
-    '''List with default visible items
-    :attr:`items` is a :class:`~kivy.properties.ListProperty` and defaults
-    to [].
-    '''
-
-    allow_custom = BooleanProperty(False)
-    '''Allow/disallow a custom item to the list
-    :attr:`allow_custom` is a :class:`~kivy.properties.BooleanProperty`
-    and defaults to False
-    '''
-
-    popup = ObjectProperty(None, allownone=True)
-    '''(internal) Used to store the current popup when it is shown.
-
-    :attr:`popup` is an :class:`~kivy.properties.ObjectProperty` and defaults
-    to None.
-    '''
-
-    def on_panel(self, instance, value):
-        if value is None:
-            return
-        self.bind(on_release=self._create_popup)
-
-    def _create_popup(self, instance):
-        # create the popup
-        content = SettingListContent(setting=self)
-        popup_width = min(0.95 * Window.width, 500)
-        popup_height = min(0.95 * Window.height, 500)
-        self.popup = popup = Popup(
-            content=content, title=self.title, size_hint=(None, None),
-            size=(popup_width, popup_height), auto_dismiss=False)
-
-        content.bind(on_apply=self._set_values, on_cancel=self.popup.dismiss)
-        selected_items = self.value.split(',')
-        # update the item list with custom values
-        for item in selected_items:
-            item = item.strip()
-            if item and not item in self.items:
-                self.items.append(item)
-        # list of items saved in the property
-        content.selected_items = selected_items
-
-        content.show_items()
-        popup.open()
-
-    def _set_values(self, *args):
-        '''Read items and save them
-        '''
-        selected_items = args[1]
-        self.value = ','.join(selected_items)
-        self.popup.dismiss()
 
 
 class SpecCodeInput(BoxLayout):
@@ -373,8 +134,8 @@ class BuildozerSpecEditor(Settings):
 
     def __init__(self, **kwargs):
         super(BuildozerSpecEditor, self).__init__(**kwargs)
-        self.register_type('dict', SpecSettingDict)
-        self.register_type('list', SpecSettingList)
+        self.register_type('dict', SettingDict)
+        self.register_type('list', SettingList)
         self.SPEC_PATH = ''
         self.proj_dir = ''
         self.config_parser = ConfigParser(name="buildozer_spec")
@@ -406,6 +167,9 @@ class BuildozerSpecEditor(Settings):
         raw_spec = SpecCodeInput(spec_path=self.SPEC_PATH)
         raw_spec.bind(on_change=self.on_spec_changed)
         self.interface.add_panel(raw_spec, "buildozer.spec", raw_spec.uid)
+
+        menu = self.interface.menu
+        menu.selected_uid = menu.buttons_layout.children[-1].uid
 
     def on_spec_changed(self, *args):
         self.load_settings(self.proj_dir)
@@ -451,3 +215,13 @@ class BuildozerSpecEditor(Settings):
             panel.add_widget(instance)
 
         return panel
+
+    def on_config_change(self, *args):
+        designer = App.get_running_app().root
+        designer.project_watcher.stop()
+
+        self.config_parser.write()
+        super(BuildozerSpecEditor, self).on_config_change(*args)
+        
+        designer.project_watcher.start_watching(
+                                            designer.project_loader.proj_dir)
