+++ designer/start_page.py
@@ -5,12 +5,10 @@ from kivy.uix.boxlayout import BoxLayout
 from kivy.uix.gridlayout import GridLayout
 from kivy.uix.button import Button
 from kivy.uix.label import Label
-from kivy.uix.popup import Popup
 from kivy.uix.scrollview import ScrollView
 from kivy.properties import StringProperty, ObjectProperty
 from kivy.metrics import pt
 from kivy.clock import Clock
-from designer.buildozer_spec_editor import BuildozerSpecEditor
 
 
 class DesignerLinkLabel(Button):
